Build scripts and template efficiency
Templatize the source dir in one pass, not a bunch of random intermediate chunks
 - Removes some redundant code and unneeded things
Add the Jenkins build scripts to the templates, so each version can be quickly tested from the console
