+++ targets/UnrealMarketplacePlugin/source/BuildAndroid.bat.ejs
@@ -1,0 +1,4 @@
+echo ========== BUILDING ANDROID ==========
+set ueVersion=<%- ueTargetVersion %>
+call "C:\Program Files\Epic Games\UE_%ueVersion%\Engine\Build\BatchFiles\RunUAT.bat" BuildPlugin -plugin="%~dp0\PlayFab\PlayFab.uplugin" -package="%~dp0\..\Build\Android" -TargetPlatforms=Android
+pause
